+++ monad-archive/src/model/mod.rs
@@ -23,7 +23,7 @@ use alloy_primitives::BlockHash;
 use alloy_rlp::{RlpDecodable, RlpEncodable};
 use enum_dispatch::enum_dispatch;
 use eyre::{OptionExt, Result};
-use monad_triedb_utils::triedb_env::{ReceiptWithLogIndex, TxEnvelopeWithSender};
+use monad_eth_types::{ReceiptWithLogIndex, TxEnvelopeWithSender};
 use serde::{Deserialize, Serialize};
 
 use crate::prelude::*;
@@ -150,6 +150,6 @@ pub struct HeaderSubset {
     pub block_timestamp: u64,
     /// Index of this transaction within the block
     pub tx_index: u64,
-    pub gas_used: u128,
+    pub gas_used: u64,
     pub base_fee_per_gas: Option<u64>,
 }
